requirements.txt
streamlit
pytesseract
opencv-python
pandas
openpyxl
pillow